Move LocalDataStore into new persistence.local package
