+++ gnd/src/main/java/com/google/android/gnd/GndApplicationModule.java
@@ -19,7 +19,7 @@ package com.google.android.gnd;
 import android.app.Application;
 import com.google.android.gms.common.GoogleApiAvailability;
 import com.google.android.gnd.inject.ActivityScoped;
-import com.google.android.gnd.repository.local.LocalDataStore;
+import com.google.android.gnd.persistence.local.LocalDataStore;
 import com.google.android.gnd.repository.local.RoomDataStore;
 import com.google.android.gnd.service.RemoteDataStore;
 import com.google.android.gnd.service.firestore.FirestoreDataStore;
